+++ bootique/src/main/java/io/bootique/Bootique.java
@@ -22,7 +22,6 @@ import java.util.Objects;
 import java.util.ServiceLoader;
 import java.util.function.Supplier;
 
-import static java.util.Collections.emptySet;
 import static java.util.stream.Collectors.joining;
 
 /**
@@ -451,7 +450,7 @@ public class Bootique {
         // is safe to do, as it won't be used until the Injector is created by the method caller.
         bqModules.add(coreModuleProvider(modulesSource).moduleBuilder().build());
 
-        BootiqueUtils.moduleProviderDependencies(builderProviders(), emptySet())
+        BootiqueUtils.moduleProviderDependencies(builderProviders())
             .forEach(p -> bqModules.add(p.moduleBuilder().build()));
 
         if (autoLoadModules) {